Drivers: platform: modify parameter type
Modify parameter type of bytes_number from uint8_t to uint16_t in
function spi_write_and_read. Modify file in includes and in all platform
files.
Uint8_t data type is not enough to represent the size of a data block
(512 bytes) for reading and writing data in a SD card over SPI.
This modification doesn't affect the existing spi drivers because in the
implementation there are no constrains for this parameter to be uint8_t

+++ include/spi.h
@@ -91,6 +91,6 @@ int32_t spi_remove(struct spi_desc *desc);
 /* Write and read data to/from SPI. */
 int32_t spi_write_and_read(struct spi_desc *desc,
 			   uint8_t *data,
-			   uint8_t bytes_number);
+			   uint16_t bytes_number);
 
 #endif // SPI_H_
